Declare DirectPublisher exchanges with passive=True
If rabbit dies, the consumer can be disconnected before the publisher
sends, and if the consumer hasn't delared the queue, the publisher's
will send a message to an exchange that's not bound to a queue, and
the message wll be lost.  Setting passive=True will cause the
publisher to fail and retry if the consumer hasn't declared the
receiving queue yet.

Closes-Bug: 1338732
Change-Id: I5ba4d311b97236d3a85a9f5badff61f12b08c12d